Estrutura do perfil completa, histórico incompleto
