--- src/Hooks/useForm.jsx
@@ -1,16 +1,0 @@
-import { useState } from "react";
-
-const useForm = (initialState) => {
-    const [form, setForm] = useState(initialState)
-
-    const onChange = (event) => {
-        const { name, value } = event.target
-        setForm({ ...form, [name]: value })
-    }
-
-    const cleanFields = () => setForm(initialState)
-
-    return { form, onChange, cleanFields }
-}
-
-export default useForm
