--- src/services/index.js
@@ -1,15 +1,0 @@
-import axios from "axios";
-import { baseURL } from "../constants";
-
-export const getProfile = (setUser, token) => {
-
-    axios.get(`${baseURL}/profile`,{
-        headers:{
-            auth: `${token}`
-        }
-    }).then((res)=>{
-        setUser(res.data.user)
-    }).catch((err)=>{
-        alert("vo vê e te aviso")
-    })
-}
